Fixed gcc warnings on 64 bit OS390 (Linux) for printf() and sscanf()
git-svn-id: svn://www.xmlblaster.org/xmlBlaster/trunk@10591 00fa6119-6a0f-0410-b6da-8b96a2b62773

+++ xmlBlaster/demo/c/socket/HelloWorld3.c
@@ -228,11 +228,11 @@ int main(int argc, char** argv)
                                              msgUnitArr->msgUnitArr[i].contentLen);
             const char *dots = (msgUnitArr->msgUnitArr[i].contentLen > 96) ?
                                  " ..." : "";
-            printf("\n[client] Received message#%u/%u:\n"
+            printf("\n[client] Received message#%lu/%lu:\n"
                      "-------------------------------------"
                      "%s\n <content>%.100s%s</content>%s\n"
                      "-------------------------------------\n",
-                     i+1, msgUnitArr->len,
+                     (unsigned long)(i+1), (unsigned long)msgUnitArr->len,
                      msgUnitArr->msgUnitArr[i].key,
                      contentStr, dots,
                      msgUnitArr->msgUnitArr[i].qos);
